Added deletion and event emitter

+++ src/App.tsx
@@ -2,26 +2,28 @@ import Devices from './components/Devices';
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator, TransitionPresets } from '@react-navigation/stack';
 import Splash from './components/Splash';
-import { MD3LightTheme, Provider, MD3Theme, Appbar, Text, IconButton } from 'react-native-paper';
+import { MD3LightTheme, Provider, MD3Theme, Appbar } from 'react-native-paper';
 import { changeBarColors } from 'react-native-immersive-bars';
 import { Image } from 'react-native';
-import { SafeAreaView, useSafeAreaInsets, SafeAreaProvider } from 'react-native-safe-area-context';
+import { useSafeAreaInsets, SafeAreaProvider } from 'react-native-safe-area-context';
 import { Device } from './objects/Device';
 import EditDevice from './components/EditDevice';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { NavigationActions } from 'react-navigation';
-import { View } from "react-native";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { DeviceManager } from './objects/DeviceManager';
+import AddDevice from './components/AddDevice';
+import SystemSetting from "react-native-system-setting";
+import { useEffect, useState } from 'react';
+import { ToastAndroid } from "react-native";
+import {NativeEventEmitter, Keyboard} from "react-native";
 const Stack = createStackNavigator();
 
 export type RootStackParamList = {
     Splash: undefined;
     Devices: undefined;
-    EditDevice: { device: Device }
+    EditDevice: { device: Device },
+    AddDevice: undefined
 };
-
-
 
 
 export type Props = NativeStackScreenProps<RootStackParamList>;
@@ -37,15 +39,33 @@ const customTheme = ({
     }
 } as MD3Theme)
 
+
+let previousVolume: number;
 function App1() {
     changeBarColors(true);
+    
+    useEffect(() => {
+        
+        const volumeListener = SystemSetting.addVolumeListener(async (data) => {
+            console.log("-------------------------------");
+            
+            console.log(data.value);
+            console.log(previousVolume);
+            new NativeEventEmitter()
+            if (!previousVolume) previousVolume = data.value;
+            if (previousVolume != data.value) {
+                await AsyncStorage.clear();
+            }
+            previousVolume = data.value;
+        });
+
+        return () => SystemSetting.removeVolumeListener(volumeListener);
+
+    }, []);
+
 
     const insets = useSafeAreaInsets();
-    (async () => {
-        await AsyncStorage.clear();
-        await DeviceManager.addDevice(new Device("device 1", "123"));
-        await DeviceManager.addDevice(new Device("device 2", "333"));
-    })();
+
     return (
         <SafeAreaProvider>
             <NavigationContainer>
@@ -64,6 +84,10 @@ function App1() {
                             ...TransitionPresets.SlideFromRightIOS, headerShown: false
                             // @ts-ignore
                         }} name="EditDevice" component={EditDevice} />
+                        <Stack.Screen options={{
+                            ...TransitionPresets.SlideFromRightIOS, headerShown: false
+                        }} name="AddDevice" component={AddDevice} />
+
 
                     </Stack.Navigator>
                 </Provider>
